SPARK-1483: Privacy plugin could freeze spark at startup -The getInstance-method of the Privacy-Plugin could freeze spark if it takes to long. This patch initializes the Privacy-Plugin in a thread. Additionally, the PrivacyManager try to discover the xmppHost, if this response take to long, the host will not be discovered at all. This patch will resolve this problem by try to discover the xmpphost again, after timeout.
SPARK-1484: Retrieving bookmarks could freeze spark at startup
-The initialization of the Bookmark-Plugin could freeze spark if it takes to long.
This patch initializes the Bookmark-Plugin in a thread.

SPARK-1498: Give error-message if download directory is invalid and user try to receive a file
-German translation

SPARK-1499: On startup, TransportUtils try to receive gateway preferences from server. If response took to long, the preferences are not retrieved at all.
-With this patch, spark try to receive the transportutils until a response from server was received.
Additionally, if gateway preferencs where accessed, spark checks if the preferences are null or not.